Refactor code. Get rid of User and Tweet classes in Service. Erase unnecessary code.
Only repository is permitted to handle User and Tweet classes, so it can not have any methods that return this classes.
Service do not have any logic, just redirect calls to repository and get information from it.
Outside of repository we can operate with only userName(which is user's identifier) and tweetId. Any other information
we are permitted to see in String format.
Delete all unnecessary code connected with Spring.